+++ src/Presentation/SmartStore.Web/Controllers/CheckoutController.cs
@@ -198,6 +198,8 @@ namespace SmartStore.Web.Controllers
 					SystemCustomerAttributeNames.OfferedShippingOptions,
 					getShippingOptionResponse.ShippingOptions,
 					_storeContext.CurrentStore.Id);
+
+				var shippingMethods = _shippingService.GetAllShippingMethods();
             
                 foreach (var shippingOption in getShippingOptionResponse.ShippingOptions)
                 {
@@ -218,7 +220,7 @@ namespace SmartStore.Web.Controllers
                     //adjust rate
                     Discount appliedDiscount = null;
                     var shippingTotal = _orderTotalCalculationService.AdjustShippingRate(
-                        shippingOption.Rate, cart, out appliedDiscount);
+						shippingOption.Rate, cart, shippingOption.Name, shippingMethods, out appliedDiscount);
 
                     decimal rateBase = _taxService.GetShippingPrice(shippingTotal, _workContext.CurrentCustomer);
                     decimal rate = _currencyService.ConvertFromPrimaryStoreCurrency(rateBase, _workContext.WorkingCurrency);