+++ src/Presentation/SmartStore.Web/Controllers/ShoppingCartController.cs
@@ -1821,6 +1821,8 @@ namespace SmartStore.Web.Controllers
                 {
                     if (getShippingOptionResponse.ShippingOptions.Count > 0)
                     {
+						var shippingMethods = _shippingService.GetAllShippingMethods();
+
                         foreach (var shippingOption in getShippingOptionResponse.ShippingOptions)
                         {
                             var soModel = new EstimateShippingModel.ShippingOptionModel()
@@ -1831,8 +1833,8 @@ namespace SmartStore.Web.Controllers
                             };
                             //calculate discounted and taxed rate
                             Discount appliedDiscount = null;
-                            decimal shippingTotal = _orderTotalCalculationService.AdjustShippingRate(shippingOption.Rate,
-                                cart, out appliedDiscount);
+                            decimal shippingTotal = _orderTotalCalculationService.AdjustShippingRate(
+								shippingOption.Rate, cart, shippingOption.Name, shippingMethods, out appliedDiscount);
 
                             decimal rateBase = _taxService.GetShippingPrice(shippingTotal, _workContext.CurrentCustomer);
                             decimal rate = _currencyService.ConvertFromPrimaryStoreCurrency(rateBase, _workContext.WorkingCurrency);