+++ changelog.md
@@ -5,8 +5,9 @@
 ###New Features###
 
 * (Developer) Admin: Implemented _entity-commit_ event message (for client EventBroker) in order to support custom data persistence in a loosely coupled manner.
-* #175: IPayment plugin: Implement deactivation of specific credit card types
+* #175 IPayment plugin: Implement deactivation of specific credit card types
 * #112 Implement 'WebApi'
+* #137 Shipping method option to ignore additional shipping charges
 
 ###Bugfixes###